moving _globalInit function to core/init.js to allow inclusion in a module, issue #2028
