+++ src/core/core.js
@@ -657,4 +657,5 @@ p5.prototype._createFriendlyGlobalFunctionBinder = function(options) {
   };
 };
 
+
 module.exports = p5;
